Update SpotifyDiscogsCollection-publishv1.py
Updated duplicate check logic. 
Removed creation of a playlist based on Spotify trackID.

Added creation of local JSON file based on Discogs collection. 
If the album is already in the JSON file, it will skip the Spotify search. 
If the album is not cached, it will search.

Printout info outlines total albums searched and what, if any, albums were added.